Removed new methods from ICameraSceneNode. Added gui sprite banks for holding bitmap fonts, window symbols, animated buttons etc. Removed IGUIFontASCII and replaced it with IGUIFontBitmap for sprite based bitmap fonts. These will be slower than normal fonts until sprite banks have proper font drawing routines. Added sprite banks to skins and allowed users to customize window symbols. Renamed CTextSceneNode2 to CBillboardTextSceneNode, large sections are currently commented out- needs updating to use the sprite bank of the bitmap font. Skins are now serializable, values are saved/loaded with the gui environment. Fixed all elements to use AbsoluteClippingRect instead of AbsoluteRect in picking. Colour select and file load now use default texts from the skin. Added a couple more operator overloads to dimension2d and position2d. Added enum attributes to GUI editor.
note: haven't changed button code to use sprite banks yet, so many buttons are currently blank. I thought it was better to commit all these changes before we get too out of sync!

git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@445 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ tools/IrrFontTool/newFontTool/CFontTool.cpp
@@ -81,7 +81,7 @@ inline int getTextureSizeFromSurfaceSize(int size)
 
 		LOGFONTW lf;
 		lf.lfFaceName[0] = L'\0';
-		lf.lfCharSet = charsets[currentCharSet];
+		lf.lfCharSet = (BYTE) charsets[currentCharSet];
 		// HRESULT hr; // no error checking(!)
 
 		// clear font list
